Reorganize misc, add MISSING for default arg values
